requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
requests==2.31.0
openai==0.28.1
PyGithub==1.59.1
python-multipart==0.0.6
jinja2==3.1.2